unittests

+++ srcs/AStar/AStar.cpp
@@ -1,6 +1,6 @@
 #include "AStar.hpp"
 
-float AStar::manhattan_metric(Point p, Point goal)
+double AStar::manhattan_metric(Point p, Point goal)
 {
     return abs(p.x - goal.x) + abs(p.y - goal.y);
 }
